+++ CHANGELOG.md
@@ -14,19 +14,26 @@
 * Add support for passing a key paths collection (`KeyPathsCollection`) when using `IRealmCollection.SubscribeForNotifications`. Passing a `KeyPathsCollection` allows to specify which changes in properties should raise a notification.
 
   A `KeyPathsCollection` can be obtained by:
-  - building it explicitly by using the method `KeyPathsCollection.Of`;
+  - building it explicitly by using the methods `KeyPathsCollection.Of` or `KeyPathsCollection.Of<T>`;
   - building it implicitly with the conversion from a `List` or array of `KeyPath` or strings;
   - getting one of the static values `Full` and `Shallow` for full and shallow notifications respectively.
+
+  A `KeyPath` can be obtained by implicit conversion from a string or built from an expression using the `KeyPath.ForExpression<T>` method.
 
   For example:
   ```csharp
   var query = realm.All<Person>();
 
+  KeyPath kp1 = "Email";
+  KeyPath kp2 = KeyPath.ForExpression<Person>(p => p.Name);
+
   KeyPathsCollection kpc;
 
   //Equivalent declarations
   kpc = KeyPathsCollection.Of("Email", "Name");
+  kpc = KeyPathsCollection.Of<Person>(p => p.Email, p => p.Name);
   kpc = new List<KeyPath> {"Email", "Name"};
+  kpc = new List<KeyPath> {kp1, kp2};
 
   query.SubscribeForNotifications(NotificationCallback, kpc);
   ```